+++ web/testing/global-setup.ts
@@ -3,7 +3,9 @@ import path from 'path'
 import * as fs from 'node:fs'
 import * as http from 'node:http'
 
-const dataFilePath = path.join(__dirname, '../public/gameData_v1.2-00.json')
+import { config } from '../src/config/config'
+
+const dataFilePath = path.join(__dirname, `../public/gameData_v${config.dataVersion}.json`)
 let gameData: string
 
 try {